Update canvas.cs
With this change I no longer need to check a list of devices to find the resolution. This works correctly with all current iOS devices in simulation. the iPhone 6 line does not return the exact resolution because of the way the hardware handles the display, but it still displays correctly. Also in my personal use I made %Resolution into $Resolution so I could determine aspect ratio to prevent stretching on all devices. But this change is too specific to my code to include here.